🐛 fix(file_controller.go): fix error message when user not found ✨ feat(file_controller.go): add endpoint to count files in tags The error message when user not found has been fixed to be more descriptive. A new endpoint has been added to count the number of files in each tag. This endpoint allows users to see how many files are associated with each tag, which can be useful for organizing and searching for files.
✨ feat(folder_controller.go): add endpoint to get folder root
A new endpoint has been added to get the root folder of a given folder. This endpoint allows users to easily navigate to the root folder of a case.

✨ feat(tag_dto.go): add TagCountDto struct
A new struct has been added to represent the count of files associated with a tag. This struct is used in the new endpoint to count files in tags.

✨ feat(tag_repository.go): add CountFilesInTags method
A

🐛 fix(file_service.go): fix typo in mimeTypeToString variable name
✨ feat(file_service.go): add caseId return value to checkPermissionAndGetCaseId function
The variable mimeTypeSTtring was renamed to mimeTypeToString to fix a typo. The checkPermission function was renamed to checkPermissionAndGetCaseId to return the caseId value. The CountFilesInTags function was added to the FileService to count the number of files in each tag. The GetFolderRoot function was added to the FolderService to get the folder hierarchy from the root folder.